rename action directory files
