--- client/actions/recipeActions.js
@@ -1,61 +1,0 @@
-import axios from 'axios';
-
-import * as actionTypes from './actionTypes';
-
-// const URL = 'https://more-recipes-25.herokuapp.com/api/v1/';
-
-const retrieveRecipesRequest = () => ({
-  type: actionTypes.RETRIEVE_RECIPES_REQUEST,
-});
-
-const retrieveRecipesSuccess = recipes => ({
-  type: actionTypes.RETRIEVE_RECIPES_SUCCESS,
-  payload: recipes,
-});
-
-const retrieveRecipesFailure = error => ({
-  type: actionTypes.RETRIEVE_RECIPES_FAILURE,
-  payload: error,
-});
-
-const fetchRecipeRequest = () => ({
-  type: actionTypes.FETCH_RECIPE_REQUEST,
-});
-
-const fetchRecipeSuccess = recipe => ({
-  type: actionTypes.FETCH_RECIPE_SUCCESS,
-  payload: recipe,
-});
-
-const fetchRecipeFailure = error => ({
-  type: actionTypes.FETCH_RECIPE_FAILURE,
-  payload: error,
-});
-
-const retrieveRecipes = () => (dispatch) => {
-  dispatch(retrieveRecipesRequest());
-  axios.get('http://127.0.0.1:3000/api/v1/recipes')
-    .then((response) => {
-      const { data } = response;
-      dispatch(retrieveRecipesSuccess(data));
-    })
-    .catch((error) => {
-      const { response: { data } } = error;
-      dispatch(retrieveRecipesFailure(data));
-    });
-};
-
-const fetchRecipe = recipeId => (dispatch) => {
-  dispatch(fetchRecipeRequest());
-  axios.get(`http://127.0.0.1:3000/api/v1/recipes/${recipeId}`)
-    .then((response) => {
-      const { data } = response;
-      dispatch(fetchRecipeSuccess(data));
-    })
-    .catch((error) => {
-      const { response: { data } } = error;
-      dispatch(fetchRecipeFailure(data));
-    });
-};
-
-export { retrieveRecipes, fetchRecipe };
